# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Outreach Server
  version: 0.1.0
  description: |
    The `/v1` surface of the outreach engine — the real API.

    **This document covers `/v1` only.** The 110 legacy endpoints served
    alongside it are a compatibility shim (`src/api/compat/`), deleted in P12,
    and are inventoried in the extraction plan's Appendix A rather than here.
    Documenting them would give them a life they are not meant to have. Every
    compat response carries `Deprecation: true` and a `Link` header naming its
    successor below.

    ## Tenancy

    Every request is tenant-scoped, and **the tenant is never read from the body
    or the query string** — it comes from the gateway-forwarded identity headers.
    A body field named `tenantId` or `medspaId` is ignored, and the MCP router
    logs a warning when it sees one.

    ## Authentication

    The gateway terminates user auth and forwards identity headers. Three
    surfaces are deliberately reachable without them, each for a stated reason:

    | Path | Why |
    |---|---|
    | `GET /metrics` | Prometheus scrapes with no headers |
    | `GET /mcp/tools`, `GET /mcp/health` | tera-orchestrator builds its tool registry at its own startup. Discovery is schema-only; every tool that touches data is authenticated. |
    | `POST /v1/webhooks/*` | a provider callback has no headers. The signature is the credential. |
...

  license:
    name: Proprietary
servers:
  - url: https://{host}/api/communication
...
      host:
...
security:
  - gatewayIdentity: []
paths:
...
  /v1/content/render:
    post:
      operationId: postContentRender
      tags: [content]
      summary: Render a template or a raw body
      description: Supply either `templateId` or `content`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                templateId: { type: string }
                content: { type: string }
                format: { $ref: '#/components/schemas/TemplateFormat' }
                packId: { type: string }
                data: { $ref: '#/components/schemas/RenderContext' }
      responses:
        '200':
          description: The rendered output.
          content:
            application/json:
              schema:
                type: object
                required: [output, format]
                properties:
                  output: { type: string }
                  format: { $ref: '#/components/schemas/TemplateFormat' }
                  warnings: { type: array, items: { type: string } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/content/generate:
    post:
      operationId: postContentGenerate
      tags: [content]
      summary: Generate a draft from a prompt pack
      description: |
        `aiConfidence` is a **deterministic composite** of things the engine can
        observe — context completeness, less a penalty per lint warning — not
        the model's self-assessment. It is a heuristic, not a probability, and
        the approval `threshold` mode that consumes it ships disabled.
      security: [{ gatewayIdentity: [outreach:send] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [promptPackKey, channel]
              properties:
                promptPackKey: { type: string }
                channel: { $ref: '#/components/schemas/Channel' }
                playbookKey: { type: string }
                playbookGoal: { type: string }
                data: { $ref: '#/components/schemas/RenderContext' }
                overrides:
                  type: object
                  properties:
                    tone: { type: string }
                    language: { type: string }
                    model: { type: string }
      responses:
        '200':
          description: The draft.
          content: { application/json: { schema: { $ref: '#/components/schemas/GeneratedDraft' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/templates:
    get:
      operationId: getTemplates
      tags: [content]
      summary: List templates
      parameters:
        - { name: channel, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: templateType, in: query, schema: { type: string } }
        - { $ref: '#/components/parameters/Limit' }
        - { name: offset, in: query, schema: { type: integer } }
      responses:
        '200':
          description: The templates.
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates: { type: array, items: { $ref: '#/components/schemas/Template' } }
                  count: { type: integer }
    post:
      operationId: postTemplates
      tags: [content]
      summary: Create a template
      security: [{ gatewayIdentity: [outreach:templates:write] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TemplateInput' } } }
      responses:
        '201':
          description: The template.
          content: { application/json: { schema: { $ref: '#/components/schemas/Template' } } }
  /v1/templates/{id}:
    get:
      operationId: getTemplatesById
      tags: [content]
      summary: Get one template
      description: Accepts a UUID or a pack key such as `medspa.reminder.sms`.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: The template.
          content: { application/json: { schema: { $ref: '#/components/schemas/Template' } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: putTemplatesById
      tags: [content]
      summary: Update a template
      security: [{ gatewayIdentity: [outreach:templates:write] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TemplateInput' } } }
      responses:
        '200':
          description: The template.
          content: { application/json: { schema: { $ref: '#/components/schemas/Template' } } }
    delete:
      operationId: deleteTemplatesById
      tags: [content]
      summary: Delete a template
      security: [{ gatewayIdentity: [outreach:templates:write] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Deleted. }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/templates/{id}/default:
    post:
      operationId: postTemplatesByIdDefault
      tags: [content]
      summary: Make this the default for its channel and category
      description: Clears the flag on its siblings, including where category is NULL.
      security: [{ gatewayIdentity: [outreach:templates:write] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: The template.
          content: { application/json: { schema: { $ref: '#/components/schemas/Template' } } }
  /v1/templates/{id}/versions:
    get:
      operationId: getTemplatesByIdVersions
      tags: [content]
      summary: Version history
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: The versions, newest first.
          content:
            application/json:
              schema:
                type: object
                properties:
                  versions: { type: array, items: { type: object, additionalProperties: true } }
  /v1/templates/{id}/render:
    post:
      operationId: postTemplatesByIdRender
      tags: [content]
      summary: Render this template
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { data: { $ref: '#/components/schemas/RenderContext' } }
      responses:
        '200':
          description: The rendered output.
          content: { application/json: { schema: { type: object, additionalProperties: true } } }

  # ── assets ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    gatewayIdentity:
      type: apiKey
      in: header
      name: x-gateway-request
      description: |
        The gateway forwards `x-gateway-request: true` plus `x-user-id`,
        `x-user-role`, `x-user-permissions` (a JSON array), `x-tenant-id` (or
        the legacy `x-medspa-id`), `x-sub-tenant-id` (or `x-location-id`) and
        `x-sender-id` (or `x-provider-id`).

        A request without `x-gateway-request` is **403**. A request with it but
        no user context is **401**. Scopes listed on an operation are the
        permissions it requires from `x-user-permissions`; the `admin` role and
        the `outreach:admin` permission bypass the check.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    BadRequest:
      description: The request did not validate.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    NotFound:
      description: No such resource in this tenant.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
        requestId: { type: string, format: uuid }
    Channel:
      type: string
      enum: [email, sms, slack, push, webhook, in_app, voice, letter]
...
    TemplateFormat:
      type: string
      enum: [TEXT, HTML, MARKDOWN, MJML]
...
    RenderContext:
      type: object
      description: |
        Namespaced deliberately. `recipient`, `sender` and `tenant` are resolved
        by the engine; `context` is whatever the caller supplied, validated
        against the playbook's data contract.
      properties:
        recipient: { type: object, additionalProperties: true }
        sender: { type: object, additionalProperties: true }
        tenant: { type: object, additionalProperties: true }
        context: { type: object, additionalProperties: true }
        message: { type: object, additionalProperties: true }
    GeneratedDraft:
      type: object
      required: [draftId, content, channel, aiConfidence]
      properties:
        draftId: { type: string, format: uuid }
        content: { type: string }
        subject: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        aiConfidence:
          type: number
          description: |
            Deterministic composite: context completeness, less 0.15 per lint
            warning. **Not** the model's self-assessment, and a heuristic rather
            than a probability.
        promptPackKey: { type: string }
        model: { type: string }
        tokensIn: { type: integer }
        tokensOut: { type: integer }
        costUsd: { type: number }
        lintWarnings: { type: array, items: { type: string } }
    TemplateInput:
      type: object
      required: [name, channel, content]
      properties:
        key: { type: string }
        name: { type: string }
        description: { type: string }
        channel: { type: string }
        subject: { type: string }
        content: { type: string }
        format: { $ref: '#/components/schemas/TemplateFormat' }
        category: { type: string }
        tags: { type: array, items: { type: string } }
    Template:
      allOf:
        - $ref: '#/components/schemas/TemplateInput'
        - type: object
          properties:
            id: { type: string, format: uuid }
            packId: { type: [string, 'null'] }
            variables: { type: array, items: { type: string } }
            status: { type: string }
            isDefault: { type: boolean }
            version: { type: integer }
            usageCount: { type: integer }
...